move pathResolver to utils package
