+++ api/batch/src/main/java/it/gov/pagopa/rtd/transaction_filter/batch/BatchExecutor.java
@@ -3,6 +3,7 @@ package it.gov.pagopa.rtd.transaction_filter.batch;
 
 import it.gov.pagopa.rtd.transaction_filter.batch.step.PanReaderStep;
 import it.gov.pagopa.rtd.transaction_filter.batch.step.TransactionFilterStep;
+import it.gov.pagopa.rtd.transaction_filter.batch.utils.PathResolver;
 import it.gov.pagopa.rtd.transaction_filter.service.StoreService;
 import java.util.Date;
 import lombok.Data;
